2286. Booking Concert Tickets in Groups, #Segment Tree
A concert hall has n rows numbered from 0 to n - 1, each with m seats, numbered from 0 to m - 1.
Support gather(int k, int maxRow) and scatter(int k, int maxRow), always search on minimal rows

+++ Source/LeetCodeAlgo/Design/NumArray_307.cs
@@ -25,7 +25,7 @@ namespace LeetCodeAlgo.Design
 
         public int SumRange(int left, int right)
         {
-            return tree.SumRange(left, right);
+            return (int)tree.SumRange(left, right);
         }
     }
 }
